backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.1.31
click==8.1.8
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==2.1.5
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-core==2.27.2
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.44.0
tomli==2.2.1
typing-extensions==4.12.2
uvicorn==0.33.0
zipp==3.20.2
